Move export wallet actions to actions/wallet.ts
